refactor: change styles for page
