+++ src/App.tsx
@@ -1,18 +1,17 @@
 import { useState } from 'react';
 
 import Header from './components/Header';
-import Main from './components/Main';
-
-import './App.css';
+import Main from './pages/MainPage/MainPage';
+import ErrorBoundary from './components/ErrorBoundary';
 
 const App = () => {
   const [searchQuery, setSearchQuery] = useState<string>('');
 
   return (
-    <>
+    <ErrorBoundary>
       <Header setSearchQuery={setSearchQuery} />
       <Main searchQuery={searchQuery} />
-    </>
+    </ErrorBoundary>
   );
 };
