+++ src/components/ErrorBoundary.tsx
@@ -21,6 +21,7 @@ class ErrorBoundary extends Component<
   };
 
   static getDerivedStateFromError() {
+    console.log('dsds');
     return { hasError: true };
   }
