+++ src/pages/MainPage/MainPage.module.css
@@ -1,0 +1,5 @@
+.main {
+  display: grid;
+  grid-template-columns: auto 300px;
+  gap: 20px;
+}
